Cleanup for HypreSemiStructuredSolver
Creates Hypre.hpp header for initalizing hypre backend for both strucutred and semi-structured solvers.
Passes variable filename to hypre print functions instead of hard-coding the output file names
