+++ cajita/src/CMakeLists.txt
@@ -44,6 +44,7 @@ set(HEADERS_PUBLIC
 
 if(Cabana_ENABLE_HYPRE)
   list(APPEND HEADERS_PUBLIC
+    Cajita_Hypre.hpp
     Cajita_HypreStructuredSolver.hpp
     Cajita_HypreSemiStructuredSolver.hpp
     )
