+++ cajita/src/Cajita_HypreStructuredSolver.hpp
@@ -21,6 +21,7 @@
 #include <Cajita_IndexSpace.hpp>
 #include <Cajita_LocalGrid.hpp>
 #include <Cajita_Types.hpp>
+#include <Cajita_Hypre.hpp>
 
 #include <HYPRE_config.h>
 #include <HYPRE_struct_ls.h>
@@ -38,60 +39,6 @@
 
 namespace Cajita
 {
-//---------------------------------------------------------------------------//
-// Hypre memory space selection. Don't compile if HYPRE wasn't configured to
-// use the device.
-// ---------------------------------------------------------------------------//
-
-//! Hypre device compatibility check.
-template <class MemorySpace>
-struct HypreIsCompatibleWithMemorySpace : std::false_type
-{
-};
-
-// FIXME: This is currently written in this structure because HYPRE only has
-// compile-time switches for backends and hence only one can be used at a
-// time. Once they have a run-time switch we can use that instead.
-#ifdef HYPRE_USING_CUDA
-#ifdef KOKKOS_ENABLE_CUDA
-#ifdef HYPRE_USING_DEVICE_MEMORY
-//! Hypre device compatibility check - CUDA memory.
-template <>
-struct HypreIsCompatibleWithMemorySpace<Kokkos::CudaSpace> : std::true_type
-{
-};
-#endif // end HYPRE_USING_DEVICE_MEMORY
-
-//! Hypre device compatibility check - CUDA UVM memory.
-#ifdef HYPRE_USING_UNIFIED_MEMORY
-template <>
-struct HypreIsCompatibleWithMemorySpace<Kokkos::CudaUVMSpace> : std::true_type
-{
-};
-#endif // end HYPRE_USING_UNIFIED_MEMORY
-#endif // end KOKKOS_ENABLE_CUDA
-#endif // end HYPRE_USING_CUDA
-
-#ifdef HYPRE_USING_HIP
-#ifdef KOKKOS_ENABLE_HIP
-//! Hypre device compatibility check - HIP memory. FIXME - make this true when
-//! the HYPRE CMake includes HIP
-template <>
-struct HypreIsCompatibleWithMemorySpace<Kokkos::ExperimentalHIPSpace>
-    : std::false_type
-{
-};
-#endif // end KOKKOS_ENABLE_HIP
-#endif // end HYPRE_USING_HIP
-
-#ifndef HYPRE_USING_GPU
-//! Hypre device compatibility check - host memory.
-template <>
-struct HypreIsCompatibleWithMemorySpace<Kokkos::HostSpace> : std::true_type
-{
-};
-#endif // end HYPRE_USING_GPU
-
 //---------------------------------------------------------------------------//
 //! Hypre structured solver interface for scalar fields.
 template <class Scalar, class EntityType, class MemorySpace>
